+++ src/main/kotlin/dev/gtnh/intellij/environment/GtnhEnvironmentCheck.kt
@@ -1,0 +1,7 @@
+package dev.gtnh.intellij.environment
+
+import com.intellij.openapi.project.Project
+
+fun interface GtnhEnvironmentCheck {
+    fun run(project: Project): GtnhEnvironmentResult
+}